Add more documentation and fix some things
