+++ boa/src/vm/call_frame.rs
@@ -15,13 +15,21 @@ pub struct CallFrame {
     pub(crate) finally_return: FinallyReturn,
     pub(crate) finally_jump: Vec<Option<u32>>,
     pub(crate) pop_on_return: usize,
+
+    // Tracks the number of environments in the current loop block.
+    // On abrupt returns this is used to decide how many environments need to be pop'ed.
     pub(crate) loop_env_stack: Vec<usize>,
+
+    // Tracks the number of environments in the current try-catch-finally block.
+    // On abrupt returns this is used to decide how many environments need to be pop'ed.
     pub(crate) try_env_stack: Vec<TryStackEntry>,
+
     pub(crate) param_count: usize,
     pub(crate) arg_count: usize,
 }
 
 impl CallFrame {
+    /// Tracks that one environment has been pushed in the current loop block.
     pub(crate) fn loop_env_stack_inc(&mut self) {
         *self
             .loop_env_stack
@@ -29,6 +37,7 @@ impl CallFrame {
             .expect("loop environment stack entry must exist") += 1;
     }
 
+    /// Tracks that one environment has been pop'ed in the current loop block.
     pub(crate) fn loop_env_stack_dec(&mut self) {
         *self
             .loop_env_stack
@@ -36,6 +45,7 @@ impl CallFrame {
             .expect("loop environment stack entry must exist") -= 1;
     }
 
+    /// Tracks that one environment has been pushed in the current try-catch-finally block.
     pub(crate) fn try_env_stack_inc(&mut self) {
         self.try_env_stack
             .last_mut()
@@ -43,6 +53,7 @@ impl CallFrame {
             .num_env += 1;
     }
 
+    /// Tracks that one environment has been pop'ed in the current try-catch-finally block.
     pub(crate) fn try_env_stack_dec(&mut self) {
         self.try_env_stack
             .last_mut()
@@ -50,6 +61,7 @@ impl CallFrame {
             .num_env -= 1;
     }
 
+    /// Tracks that one loop has started in the current try-catch-finally block.
     pub(crate) fn try_env_stack_loop_inc(&mut self) {
         self.try_env_stack
             .last_mut()
@@ -57,6 +69,7 @@ impl CallFrame {
             .num_loop_stack_entries += 1;
     }
 
+    /// Tracks that one loop has finished in the current try-catch-finally block.
     pub(crate) fn try_env_stack_loop_dec(&mut self) {
         self.try_env_stack
             .last_mut()
@@ -65,6 +78,10 @@ impl CallFrame {
     }
 }
 
+/// Tracks the number of environments in the current try-catch-finally block.
+///
+/// Because of the interactions between loops and try-catch-finally blocks,
+/// the number of loop blocks in the try-catch-finally block also needs to be tracked.
 #[derive(Copy, Clone, Debug)]
 pub(crate) struct TryStackEntry {
     pub(crate) num_env: usize,
